audio2kafka/requirements.txt
cffi==1.17.1
kafka-python-ng==2.2.3
numpy==2.1.2
pycparser==2.22
six==1.16.0
sounddevice==0.5.1
